Reorganize and modularize Analysis
